test: use int64 when testing formatSecs

+++ internal/templates/templates_test.go
@@ -6,7 +6,7 @@ import (
 
 func TestFormatSecs(t *testing.T) {
 	tests := []struct {
-		input          uint32
+		input          int64
 		expectedOutput string
 	}{
 		{0, "0s"},
